OBPIH-7519 Improvements after review
