+++ grails-app/services/org/pih/warehouse/inventory/RecordStockProductInventoryTransactionService.groovy
@@ -1,11 +1,13 @@
 package org.pih.warehouse.inventory
 
+import grails.validation.ValidationException
 import org.pih.warehouse.api.AvailableItem
 import org.pih.warehouse.core.ConfigService
+import org.pih.warehouse.auth.AuthService
 import org.pih.warehouse.core.Constants
 
 import grails.gorm.transactions.Transactional
-
+import org.pih.warehouse.core.Location
 import org.pih.warehouse.inventory.product.availability.AvailableItemKey
 import org.pih.warehouse.inventory.product.availability.AvailableItemMap
 
@@ -34,6 +36,10 @@ class RecordStockProductInventoryTransactionService extends ProductInventoryTran
                 transactionSource: transactionSource
         )
         transaction.transactionNumber = inventoryService.generateTransactionNumber(transaction)
+        // If baseline was not created just before the adjustment, it means we have not created transaction source instance yet
+        if (!transactionSource) {
+            setSourceObject(transaction, recordInventoryCommand)
+        }
         return transaction
     }
 
@@ -84,9 +90,21 @@ class RecordStockProductInventoryTransactionService extends ProductInventoryTran
                 )}
     }
 
+    TransactionSource createRecordStockTransactionSource(Location location) {
+        TransactionSource transactionSource = new TransactionSource(
+                transactionAction: TransactionAction.RECORD_STOCK,
+                origin: location,
+                destination: location,
+        )
+        if (!transactionSource.validate()) {
+            throw new ValidationException("Invalid transaction source", transactionSource.errors)
+        }
+        return transactionSource.save()
+    }
 
     @Override
-    void setSourceObject(Transaction transaction, RecordInventoryCommand transactionSource) {
-        // Record stock has no source object.
+    void setSourceObject(Transaction transaction, RecordInventoryCommand sourceObject) {
+        TransactionSource transactionSource = createRecordStockTransactionSource(sourceObject.inventory.warehouse)
+        transaction.transactionSource = transactionSource
     }
 }
