+++ src/main/groovy/org/pih/warehouse/inventory/InventoryBaselineTransactionCommand.groovy
@@ -1,12 +1,11 @@
 package org.pih.warehouse.inventory
 
-import grails.validation.Validateable
 import org.pih.warehouse.core.Location
 import org.pih.warehouse.inventory.product.availability.AvailableItemKey
 import org.pih.warehouse.inventory.product.availability.AvailableItemMap
 import org.pih.warehouse.product.Product
 
-class InventoryBaselineTransactionCommand<T> implements Validateable {
+class InventoryBaselineTransactionCommand<T> {
     Location facility
     T sourceObject
     Collection<Product> products
@@ -16,15 +15,4 @@ class InventoryBaselineTransactionCommand<T> implements Validateable {
     Map<AvailableItemKey, String> transactionEntriesComments = [:]
     Boolean validateTransactionDates = true
     Boolean disableRefresh = false
-    TransactionSource transactionSource
-
-    static constraints = {
-        sourceObject(nullable: true)
-        availableItems(nullable: true)
-        transactionDate(nullable: true)
-        comment(nullable: true)
-        transactionEntriesComments(nullable: true)
-        // TODO: (?) should probably become required in the future if we cover all places where transaction sources should be created
-        transactionSource(nullable: true)
-    }
 }
